Removed BigInterger

+++ Archipelago.Gifting.Net/Archipelago.Gifting.Net/Gifts/Versions/Current/Gift.cs
@@ -14,7 +14,7 @@ namespace Archipelago.Gifting.Net.Gifts.Versions.Current
         /// This is the value per item, in the standard Archipelago currency.
         /// This is not the value of the whole gift. The value of the gift is the value of the item, multiplied by the amount of the item
         /// </summary>
-        public BigInteger ItemValue { get; set; }
+        public int ItemValue { get; set; }
         public GiftTrait[] Traits { get; set; }
         public int SenderSlot { get; set; }
         public int ReceiverSlot { get; set; }
@@ -26,7 +26,7 @@ namespace Archipelago.Gifting.Net.Gifts.Versions.Current
         {
         }
 
-        public Gift(string itemName, int amount, BigInteger itemValue, GiftTrait[] traits, int senderSlot, int receiverSlot, int senderTeam, int receiverTeam)
+        public Gift(string itemName, int amount, int itemValue, GiftTrait[] traits, int senderSlot, int receiverSlot, int senderTeam, int receiverTeam)
         {
             ID = Guid.NewGuid().ToString();
             ItemName = itemName;
